adding multiple features mostly in legal moves
things added:

today I added all kinds of updates to the game, mostly on the legal side. The game is now capable of making sure that you can only move if it runs through a series of checks legally. These checks include whether or not the piece is moving to the correct square and if it is within the reach of the checker piece.

the pieces have an argument in the checkIfLegal() function which is called killException.

killException will run it through a separate legal check and this will run it as though it is attempting to capture another piece.

currently, there are no real bugs in the code because it is very simple.

Next, I will add further rules on capturing pieces and then hopefully we can work our way up to having king pieces by the end of Monday.