Added UI improvements
added aesthetic improvements to fix up the game. Next is to add a function for determining a win/loss as well as king pieces that can move back.

there is also a minor bug in the spaceLimiter function which is allowing pieces to move left and backwards, which will need to be addressed as well.