major changes to functionality
- found the bug for the spaceLimiter. More information about that at the end.
- added code to determine if a player has won, lost, or drawed. Next need to add the UI to show it, as well as fix a few bugs. More about but at the end.

spaceLimiter Bugs:
default action for the spaceLimiter was True, and it would look for reasons where it should return false. Instead of doing this, I flipped it, making the default false, and let the spaceLimiter look for reasons to flip it to true. This forces it to be a lot more strict about the movements, thus fixing the bug. I also cleaned up the code as there was lots of redundancies.

Draw/Win Bugs:
currently when the code runs it shows the amount of possible moves for the last movement, rather than the new one. This is a bug because when there are no more possible moves it will wait for you to input a move before telling you, which makes it impossible. To fix this I will have to move that to the beginning of the code somehow.